requirements.txt
faker==26.0.0
kafka-python-ng==2.2.2
python-dotenv==1.0.1
pydantic_settings==2.12.0
pydantic==2.10.6
psycopg2-binary==2.9.10